refactor cose serializers
