+++ indispensable-cosef/src/commonMain/kotlin/at/asitplus/signum/indispensable/cosef/CoseSerializer.kt
@@ -3,6 +3,9 @@ package at.asitplus.signum.indispensable.cosef
 import at.asitplus.signum.indispensable.CryptoSignature
 import at.asitplus.signum.indispensable.SignatureAlgorithm
 import at.asitplus.signum.indispensable.X509SignatureAlgorithm
+import at.asitplus.signum.indispensable.cosef.CosePayloadHelper.toHeader
+import at.asitplus.signum.indispensable.cosef.CosePayloadHelper.toNullablePayload
+import at.asitplus.signum.indispensable.cosef.CosePayloadHelper.toSignature
 import at.asitplus.signum.indispensable.cosef.io.ByteStringWrapperSerializer
 import at.asitplus.signum.indispensable.cosef.io.coseCompliantSerializer
 import at.asitplus.signum.indispensable.io.Base64Strict
@@ -51,16 +54,56 @@ class CoseSignedSerializer<P : Any?>(
         return CoseSigned<P>(
             protectedHeader = protectedHeader,
             unprotectedHeader = wire.unprotectedHeader,
-            payload = wire.payload.toNullablePayload(),
+            payload = wire.payload.toNullablePayload(parameterSerializer),
             signature = signature,
             wireFormat = wire
         )
     }
+}
 
-    private fun ByteArray.toHeader(): CoseHeader =
-        coseCompliantSerializer.decodeFromByteArray(CoseHeader.serializer(), this)
+class CoseMacSerializer<P : Any?>(
+    private val parameterSerializer: KSerializer<P>,
+) : KSerializer<CoseMac<P>> {
 
-    private fun ByteArray.toSignature(
+    override val descriptor: SerialDescriptor =
+        PrimitiveSerialDescriptor("CoseMac", PrimitiveKind.BYTE)
+
+    override fun serialize(encoder: Encoder, value: CoseMac<P>) {
+        if (encoder is CborEncoder) {
+            encoder.encodeSerializableValue(CoseBytes.serializer(), value.wireFormat)
+        } else {
+            val bytes = coseCompliantSerializer.encodeToByteArray(value.wireFormat)
+            encoder.encodeString(bytes.encodeToString(Base64Strict))
+        }
+    }
+
+    override fun deserialize(decoder: Decoder): CoseMac<P> {
+        val wire = if (decoder is JsonDecoder) decoder.decodeString().decodeToByteArray(Base64())
+            .let { bytes ->
+                coseCompliantSerializer.decodeFromByteArray(CoseBytes.serializer(), bytes)
+            } else decoder.decodeSerializableValue(
+            CoseBytes.serializer()
+        )
+        val protectedHeader = wire.protectedHeader.toHeader()
+        val tag = wire.rawAuthBytes
+        return CoseMac<P>(
+            protectedHeader = protectedHeader,
+            unprotectedHeader = wire.unprotectedHeader,
+            payload = wire.payload.toNullablePayload(parameterSerializer),
+            tag = tag,
+            wireFormat = wire
+        )
+    }
+}
+
+private object CosePayloadHelper {
+
+    fun <P : Any?> ByteArray?.toNullablePayload(serializer: KSerializer<P>): P? = when {
+        this == null || this.isEmpty() -> null
+        else -> toTypedPayload(serializer)
+    }
+
+    fun ByteArray.toSignature(
         protectedHeader: CoseHeader,
         unprotectedHeader: CoseHeader?,
     ): CryptoSignature.RawByteEncodable =
@@ -68,27 +111,27 @@ class CoseSignedSerializer<P : Any?>(
             CryptoSignature.EC.fromRawBytes(this)
         else CryptoSignature.RSA(this)
 
-    private fun ByteArray?.toNullablePayload(): P? = when (this) {
-        null -> null
-        else -> if (this.isEmpty()) null else toTypedPayload()
-    }
 
-    private fun ByteArray.toTypedPayload(): P =
-        if (parameterSerializer == ByteArraySerializer()) {
+    fun <P : Any?> ByteArray.toTypedPayload(serializer: KSerializer<P>): P =
+        if (serializer == ByteArraySerializer()) {
             @Suppress("UNCHECKED_CAST")
             (this as P)
         } else {
-            runCatching { fromBytes() }
-                .getOrElse { fromByteStringWrapper() }
-            // if it still fails, the input is not valid
+            runCatching { fromBytes(serializer) }
+                .getOrElse { fromByteStringWrapper(serializer) }
         }
 
-    private fun ByteArray.fromBytes(): P =
-        coseCompliantSerializer.decodeFromByteArray(parameterSerializer, this)
+    private fun <P : Any?> ByteArray.fromBytes(serializer: KSerializer<P>): P =
+        coseCompliantSerializer.decodeFromByteArray(serializer, this)
 
-    private fun ByteArray.fromByteStringWrapper(): P =
-        coseCompliantSerializer.decodeFromByteArray(ByteStringWrapperSerializer(parameterSerializer), this).value
+    private fun <P : Any?> ByteArray.fromByteStringWrapper(serializer: KSerializer<P>): P =
+        coseCompliantSerializer.decodeFromByteArray(
+            ByteStringWrapperSerializer(serializer),
+            this
+        ).value
 
+    fun ByteArray.toHeader(): CoseHeader =
+        coseCompliantSerializer.decodeFromByteArray(CoseHeader.serializer(), this)
 }
 
 private fun CoseHeader.usesEC(): Boolean? = when (algorithm) {
@@ -98,3 +141,5 @@ private fun CoseHeader.usesEC(): Boolean? = when (algorithm) {
     is CoseAlgorithm.Signature -> (algorithm.algorithm is SignatureAlgorithm.ECDSA)
     else -> false
 }
+
+
